Finished implementing User DB and updating Navbar for Logout
