--- Second_Iteration/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-  "python.pythonPath": "TeamBestRanch.github.io\\First_Iteration\\backend\\FirstIteration_env\\Scripts\\python.exe"
-}
