+++ Second_Iteration/LIS_app/forms.py
@@ -1,7 +1,7 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField, BooleanField
 from wtforms.validators import DataRequired, Length, Email, EqualTo, ValidationError
-# from app import User NEED TO FIX  CAN'T IMPORT FROM APP RIGHT NOW
+from LIS_app.database import User
 
 class RegistrationForm(FlaskForm):
     fname = StringField('First Name', validators=[
@@ -15,10 +15,13 @@ class RegistrationForm(FlaskForm):
         'Confirm Password', validators=[DataRequired(), EqualTo('password')])
     submit_signup = SubmitField('Sign Up')
     
-    def validate_field(self, email):
+        
+    # Custom validation for users
+    def validate_email(self, email):
         user = User.query.filter_by(email = email.data).all()
         if user:
             raise ValidationError('Account for email already exists!')
+        
 
 
 class LoginForm(FlaskForm):
