+++ Second_Iteration/run.py
@@ -1,0 +1,5 @@
+from LIS_app import app
+
+# Runs the Application
+if(__name__) == '__main__':
+    app.run(debug=True)
